refactor: add stack params list
