+++ src/types/index.ts
@@ -1,1 +1,2 @@
 export * from './bottomSheet';
+export * from './stackParams';
