+++ src/types/stackParams/auth.ts
@@ -1,0 +1,7 @@
+export type AuthStackParamList = {
+  AuthMain: undefined;
+  Login: undefined;
+  Register: undefined;
+  VerifyCode: undefined;
+  Verify: undefined;
+};
