+++ src/types/stackParams/community.ts
@@ -1,0 +1,7 @@
+export type CommunityStackParamList = {
+  CommunityMain: undefined;
+  CommunityPostDetail: { id: number };
+  CommunityCreatePost: undefined;
+  CommunityVisibleType: undefined;
+  CommunityAnonymitySettings: undefined;
+};
