+++ src/types/stackParams/eoullim.ts
@@ -1,0 +1,7 @@
+export type EoullimStackParamList = {
+  EoullimMain: undefined;
+  EoullimVote: undefined;
+  EoullimTimeTable: undefined;
+  EoullimRaffle: undefined;
+  EoullimStatus: undefined;
+};
