+++ src/types/stackParams/hanowl.ts
@@ -1,0 +1,7 @@
+export type HanowlStackParamList = {
+  HanowlSelectTeam: undefined;
+  HanowlApplyDetails: undefined;
+  HanowlFinalConfirm: undefined;
+  HanowlConfirm: undefined;
+  HanowlMain: undefined;
+};
