+++ src/types/stackParams/hanumPay.ts
@@ -1,0 +1,6 @@
+export type HanumPayStackParamList = {
+  HanumPayMain: undefined;
+  HanumPayQR: undefined;
+  HanumPayStatus: undefined;
+  HanumPay: undefined;
+};
