+++ src/types/stackParams/main.ts
@@ -1,0 +1,8 @@
+export type MainStackParamList = {
+  Main: undefined;
+  Schedule: undefined;
+  TimeTable: undefined;
+  Home: undefined;
+  UserInfo: undefined;
+  WebView: undefined;
+};
